Merge Azcopy Dev and xdatamove main

+++ e2etest/declarativeResourceManagers.go
@@ -27,9 +27,9 @@ import (
 	"runtime"
 	"strings"
 
-	"github.com/Azure/azure-storage-azcopy/v10/common"
 	"github.com/Azure/azure-storage-blob-go/azblob"
 	"github.com/Azure/azure-storage-file-go/azfile"
+	"github.com/shubham808/azure-storage-azcopy/v10/common"
 )
 
 func assertNoStripTopDir(stripTopDir bool) {
@@ -44,6 +44,7 @@ type downloadContentOptions struct {
 	downloadFileContentOptions
 }
 
+// nolint
 type downloadBlobContentOptions struct {
 	containerURL azblob.ContainerURL
 	cpkInfo      common.CpkInfo
@@ -95,13 +96,17 @@ type resourceManager interface {
 	createSourceSnapshot(a asserter)
 }
 
-///////////////
+// /////////////
 
 type resourceLocal struct {
 	dirPath string
 }
 
 func (r *resourceLocal) createLocation(a asserter, s *scenario) {
+	if r.dirPath == common.Dev_Null {
+		return
+	}
+
 	r.dirPath = TestResourceFactory{}.CreateLocalDirectory(a)
 	if s.GetModifiableParameters().relativeSourcePath != "" {
 		r.appendSourcePath(s.GetModifiableParameters().relativeSourcePath, true)
@@ -109,6 +114,10 @@ func (r *resourceLocal) createLocation(a asserter, s *scenario) {
 }
 
 func (r *resourceLocal) createFiles(a asserter, s *scenario, isSource bool) {
+	if r.dirPath == common.Dev_Null {
+		return
+	}
+
 	scenarioHelper{}.generateLocalFilesFromList(a, &generateLocalFilesFromList{
 		dirPath: r.dirPath,
 		generateFromListOptions: generateFromListOptions{
@@ -119,6 +128,10 @@ func (r *resourceLocal) createFiles(a asserter, s *scenario, isSource bool) {
 }
 
 func (r *resourceLocal) createFile(a asserter, o *testObject, s *scenario, isSource bool) {
+	if r.dirPath == common.Dev_Null {
+		return
+	}
+
 	scenarioHelper{}.generateLocalFilesFromList(a, &generateLocalFilesFromList{
 		dirPath: r.dirPath,
 		generateFromListOptions: generateFromListOptions{
@@ -129,12 +142,20 @@ func (r *resourceLocal) createFile(a asserter, o *testObject, s *scenario, isSou
 }
 
 func (r *resourceLocal) cleanup(_ asserter) {
+	if r.dirPath == common.Dev_Null {
+		return
+	}
+
 	if r.dirPath != "" {
 		_ = os.RemoveAll(r.dirPath)
 	}
 }
 
 func (r *resourceLocal) getParam(stripTopDir bool, withSas bool, withFile string) string {
+	if r.dirPath == common.Dev_Null {
+		return common.Dev_Null
+	}
+
 	if !stripTopDir {
 		if withFile != "" {
 			p := path.Join(r.dirPath, withFile)
@@ -164,6 +185,10 @@ func (r *resourceLocal) appendSourcePath(filePath string, _ bool) {
 }
 
 func (r *resourceLocal) getAllProperties(a asserter) map[string]*objectProperties {
+	if r.dirPath == common.Dev_Null {
+		return make(map[string]*objectProperties)
+	}
+
 	return scenarioHelper{}.enumerateLocalProperties(a, r.dirPath)
 }
 
@@ -175,7 +200,7 @@ func (r *resourceLocal) createSourceSnapshot(a asserter) {
 	panic("Not Implemented")
 }
 
-///////
+// /////
 
 type resourceBlobContainer struct {
 	accountType  AccountType
@@ -199,12 +224,15 @@ func (r *resourceBlobContainer) createFiles(a asserter, s *scenario, isSource bo
 		generateFromListOptions: generateFromListOptions{
 			fs:          s.fs.allObjects(isSource),
 			defaultSize: s.fs.defaultSize,
-			accountType: s.accountType,
+			accountType: s.srcAccountType,
 		},
 	}
 	if s.fromTo.IsDownload() {
 		options.cpkInfo = common.GetCpkInfo(s.p.cpkByValue)
 		options.cpkScopeInfo = common.GetCpkScopeInfo(s.p.cpkByName)
+	}
+	if isSource {
+		options.accessTier = s.p.accessTier
 	}
 	scenarioHelper{}.generateBlobsFromList(a, options)
 }
@@ -278,7 +306,7 @@ func (r *resourceBlobContainer) createSourceSnapshot(a asserter) {
 	panic("Not Implemented")
 }
 
-/////
+// ///
 
 type resourceAzureFileShare struct {
 	accountType AccountType
@@ -374,7 +402,7 @@ func (r *resourceAzureFileShare) createSourceSnapshot(a asserter) {
 	r.snapshotID = TestResourceFactory{}.CreateNewFileShareSnapshot(a, *r.shareURL)
 }
 
-////
+// //
 
 type resourceDummy struct{}
